// src/app/interfaces/prengutas.interface.ts
// Generated by https://quicktype.io

export interface Preguntas {
  response_code: number;
  results: Pregunta[];
}

export interface Pregunta {
  category: string;
  type: Type;
  difficulty: Difficulty;
  question: string;
  correct_answer: string;
  incorrect_answers: string[];
  group: string;
}

export enum Difficulty {
  Easy = "easy",
  Hard = "hard",
  Medium = "medium",
}

export enum Type {
  Boolean = "boolean",
  Multiple = "multiple",
}